[Feature] Introduce safe mode for starrocks cluster (#21855)
* [Feature] Introduce safe mode for starrocks cluster

In this pr, we introduced safe mode for StarRocks cluster.
- use be.getMaxDiskUsedPct() and does not consider trash now
- introduce a fe config enable_safe_mode_check(default value is true)

The detailed explanation is as follows:

When the usage of one disk exceeds Config.safe_mode_check_disk_ratio(default 0.8), it enters safemode and does not allow any new load transactions to come in. At this point, there are two situations for the ongoing load:
1. For stream load, because the transaction of stream load cannot be particularly large (usually less than 100M), so  of the disk capacity is enough. So just wait for it to complete without abort.
2. Notice that flink/datax/routine load are all using stream load ,so they are handled the same way as stream load
3. For broker load, because the amount of transaction data of broker load is unpredictable, maybe 1-Config.safe_mode_check_disk_ratio(default 0.2) is not enough, so it needs to be killed. However, considering that the broker load may be imported "a little bit", if it is killed at this time, all previous efforts may be wasted, so it is allowed to run for a while, if it really exceeds Config.safe_mode_abort_txn_disk_ratio(default is 0.9), It's not too late for us to kill it, so mark it as pending_cancel at Config.safe_mode_check_disk_ratio.
4. Since the disk usage may also be reduced (delete data, compaction, etc.), on the one hand, new load transactions should not be allowed to enter, and on the other hand, it should be as friendly as possible to existing load transactions.

Other design tradeoffs:

We did not use the editlog to store the safemode information, but kept it directly in the memory of the fe leader node. Because the introduction of editlog may face compatibility issues and some other corner cases related to fe metadata.

When we keep the safe mode state in the memory, if the leader is switched or restarted, startLeaderOnlyDaemonThreads will first execute a heartbeat task to obtain the status of the be node, and then execute the logic of the safe mode checker task. This ensures that there are no load jobs even if we restart fe or leader switched.

If heartbeat task or safe mode checker was not executed, theoretically load manager will also not be available, so there will be no load jobs too.

Signed-off-by: caneGuy <zhoukang199191@gmail.com>